+++ src/main/java/com/vaempunk/stafftool/repository/EmployeeRepository.java
@@ -6,11 +6,11 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.vaempunk.stafftool.entity.Employee;
 
-public interface EmployeeRepository extends CrudRepository<Employee, Integer> {
+public interface EmployeeRepository extends CrudRepository<Employee, Long> {
 
     public List<Employee> findAll();
     public boolean existsByEmail(String email);
     public boolean existsByPhoneNumber(String phoneNumber);
-    public boolean existsByEmailAndIdNot(String email, Integer id);
-    public boolean existsByPhoneNumberAndIdNot(String phoneNumber, Integer id);
+    public boolean existsByEmailAndIdNot(String email, long id);
+    public boolean existsByPhoneNumberAndIdNot(String phoneNumber, long id);
 }
